api_rag/requirements.txt
fastapi
uvicorn
pydantic
langchain-google-genai
python-dotenv
chromadb
transformers
torch
numpy